modified:   examples/DGmethods/ex_001_dycoms.jl modified:   test/DGmethods/compressible_Navier_Stokes/density_current-model.jl modified:   test/DGmethods/compressible_Navier_Stokes/rising_bubble-model.jl

+++ examples/DGmethods/ex_001_dycoms.jl
@@ -198,7 +198,7 @@ let
   end
   @testset "$(@__FILE__)" for ArrayType in ArrayTypes
     # Problem type
-    DT = Float64
+    DT = Float32
     # DG polynomial order 
     polynomialorder = 4
     # User specified grid spacing
